Added anti-aliasing to raytracer

+++ Source/GOATracer/Raytracer/Camera.cs
@@ -47,8 +47,6 @@ namespace GOATracer.Raytracer
             _rotation = rotation;
         }
 
-        // --- MANUAL RIGHT-HANDED MATRICES (AS PER REQUIREMENT) ---
-
         /// <summary>
         /// Creates a Right-Handed LookAt matrix (like OpenGL's gluLookAt).
         /// </summary>
@@ -81,8 +79,6 @@ namespace GOATracer.Raytracer
             );
         }
 
-        // --- MATRIX GETTERS (NOW USING CUSTOM METHODS) ---
-
         private Matrix4x4 GetViewMatrix()
         {
             Vector3 forward = _direction;
@@ -106,9 +102,8 @@ namespace GOATracer.Raytracer
             return CreateCustomPerspectiveFieldOfViewRightHanded(fovRad, aspectRatio, nearPlane, farPlane);
         }
 
-        // --- GETRAYDIRECTION (REMAINS THE SAME) ---
 
-        public Vector3 GetRayDirection(int x, int y, int imageWidth, int imageHeight)
+        public Vector3 GetRayDirection(double x, double y, int imageWidth, int imageHeight)
         {
             float aspectRatio = (float)imageWidth / imageHeight;
             Matrix4x4 viewMatrix = GetViewMatrix();
